Use commander for CLI interface.
We looked at Thor and GLI as well. We've had poor experiences with Thor
in the past so that was taken off the list. GLI was next up. We decided
it was a bit too heavy weight for us. We would have ended up moving lots
of parts around to namespace things cleanly.

At this point we considered how we wanted things to work. We spiked a
simple interface (ended up being a DSL). The resulting code looked very
much like how commander is setup. Instead of re-inventing the wheel,
just use commander for the CLI.